backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django
django-cors-headers==3.11.0
djangorestframework
djangorestframework-simplejwt==5.0.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
django-import-export
django-crispy-forms
django-mathfilters
django-ckeditor
django-ckeditor-5
django-taggit
django-jazzmin
setuptools
crewai
crewai-tools
langchain
langchain-tools
shortuuid
django-shortuuidfield